requirements.txt
requests==2.31.0
pyTelegramBotAPI==4.14.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1